requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.3.1+cpu
torchvision==0.4.2+cpu
Cython==0.29.14
numpy==1.17.3
scipy==1.1.0
dominate==2.4.0
visdom==0.1.8.9
lmdb==0.98
Pillow==6.2.1
nltk>=3.4.5
natsort==6.0.0
Flask==1.1.1
scikit_learn==0.21.3
flask-restful==0.3.7
gunicorn==19.9.0
Werkzeug==0.16.0
opencv-python==3.4.2.17
scikit-image==0.14.2
Flask-Cors==3.0.8
tensorflow==2.0.0		
-e git://github.com/faustomorales/keras-ocr.git#egg=keras-ocr